parte server da non toccare per evitare conflitti
