+++ Back-end/moduli/mod1.py
@@ -1,0 +1,18 @@
+from moduli import queryLib
+queryLib.connetti()
+
+
+
+def check_get(path):
+    pass
+
+
+def check_post(path,clientchoice):
+    pass
+
+
+if __name__ =="__main__":
+    risultato = queryLib.execute('SELECT * FROM classi')
+    for tupla in risultato:
+        print(tupla[0])
+queryLib.disconnetti()
