+++ Back-end/moduli/modulo_test/script.py
@@ -1,16 +1,7 @@
 def check_get(path):
-    if path.endswith("get_generica"):
-        return funz()
+    if path.endswith('get_generica'):
+        print('ciao')
 
 def check_post(path, client_choice):
-    if path.endswith("post_generica"):
-        return funz222(client_choice)
-
-def funz():
-    print("ciao")
-
-def funz222(client_choice):
-    nome = client_choice.get("nome", "")
-    cognome = client_choice.get("cognome", "")
-
-    print("ciao" + nome + cognome)
+    if path.endswith('post_generica'):
+        print(client_choice)
